NickServer1
The server files I will be demoing with on 3-8-2018. It was made using the "Flask Mega Tutorial" which you can find on google. Not designed to work with game yet, just a standalone server project. Figured I might as well post it incase you want to see it. To run the server, make sure you have flask installed. type "set FLASK_APP=nickserver.py" into the terminal. This will set the environmental variable to the right file. Also, that's the command on Windows only. I think mac and Linux uses "export FLASK_APP=nickserver.py" but im not sure. Once you've done that, type "flask run" into the terminal, and it should be running.